refactor(GraphQL): More PR feedback work
